+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to the Zowe CLI package will be documented in this file.
 ## Recent Changes
 
 - Added new method "dataSetLike(session, dataSetName, options)" to `Create` class, for use when creating a dataset with parameters like another data set
+- Fix compilation error
+
 ## `6.19.0`
 
 - Add CLI command to delete migrated data sets `zowe zos-files delete migrated-data-sets`.