+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to the Zowe CLI package will be documented in this file.
 ## Recent Changes
 
 - Fix compilation error
+- Update hsmCommands to have a base options handler
 
 ## `6.19.0`