+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 All notable changes to the Zowe CLI package will be documented in this file.
 
-## Recent Changes
+## `6.19.1`
 
+- Update Imperative version
 - Fix compilation error
 - Update hsmCommands to have a base options handler